refactor(persistency): move export/import_state to CoreComponent
Relocate export_state/import_state from CoreDetector to CoreComponent so
parsers and aggregators can reuse them. Return None / no-op instead of
raising when no persistency is configured, and keep the saver lock that
guards against the background save timer.
